requirements.txt
aiohttp==3.6.2
alpaca-trade-api==0.48
alpha-vantage==2.2.0
astroid==2.4.1
async-timeout==3.0.1
asyncio-nats-client==0.10.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Deprecated==1.2.10
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.18.4
pandas==1.0.3
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
tabulate==0.8.7
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.3
websocket-client==0.57.0
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
